Refactor: Image 파일명 및 변수명 수정
- Img 파일명 및 변수명 Image로 수정
