+++ src/pages/Message/SendMessage.jsx
@@ -5,7 +5,7 @@ import SendMessageForm from './SendMessageForm';
 import ToggleButton from '../../components/Button/ToggleButton';
 import SubmitButton from '../../components/Button/SubmitButton';
 import Editor from '../../components/Text/Editor';
-import ProfileImgMain from '../../components/ProfileImg/ProfileImgMain';
+import ProfileImagesMain from '../../components/ProfileImages/ProfileImagesMain';
 import getProfileImages from '../../apis/profileApis';
 
 const relationship = ['지인', '친구', '동료', '가족'];
@@ -35,7 +35,7 @@ export default function SendMessage() {
       </div>
       <div className="MessagePage__profileImg">
         <Heading>프로필 이미지</Heading>
-        <ProfileImgMain profileImages={profileImages} />
+        <ProfileImagesMain profileImages={profileImages} />
       </div>
       <div className="MessagePage__relationship">
         <Heading>상대와의 관계</Heading>
